requirements.txt

fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
slowapi
psycopg2-binary
alembic
